test: move call arg-construction tests into test/unit/commands
The 11 test_call*_arg_construction_* cases in call_test.rb stub
send_command and never touch the network, so they belong in test/unit
rather than mixed into the standalone/cluster integration test classes.

capture_call_args/capture_call_v_args previously stubbed send_command on
`r`, the live TCP-connected client from Helper::Generic#setup. Replaced
with an unconnected_client fixture that stubs the FFI connect call
(create_client_from_uri/free_connection_response), matching the technique
in test/unit/connection_config_test.rb, so the suite runs with no server.
